move Api.MessageHandling namespace into Api
