--- Assets/Scripts/Api/MessageHandling.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 16a8c9ab997143e48a09a18cc14fc472
-timeCreated: 1677883256
