+++ Assets/Scripts/Api/MessageSender.cs
@@ -9,7 +9,7 @@ using Immerse.BfhClient.Api.Messages;
 using Newtonsoft.Json.Linq;
 using UnityEngine;
 
-namespace Immerse.BfhClient.Api.MessageHandling
+namespace Immerse.BfhClient.Api
 {
     /// <summary>
     /// Handles sending messages through the WebSocket connection to the game server.
